Tests/LibWeb/Text/expected/css/CSSPropertyRule-properties-readonly.txt

@property rule syntax value: *
@property rule syntax value: *
@property rule inherits value: false
@property rule inherits value: false
@property rule initialValue value: blue
@property rule initialValue value: blue
